+++ packages/expo-brownfield/plugin/src/ios/plugins/withPodfilePlugin.ts
@@ -1,13 +1,10 @@
 import { type ConfigPlugin, withPodfile } from 'expo/config-plugins';
 
 import type { PluginConfig } from '../types';
-import { addNewPodsTarget, addPrebuildsEnvShim, addPrebuiltSettings } from '../utils';
+import { addNewPodsTarget, addPrebuiltSettings } from '../utils';
 
 const withPodfilePlugin: ConfigPlugin<PluginConfig> = (config, pluginConfig) => {
   return withPodfile(config, (config) => {
-    if (pluginConfig.usePrebuilds) {
-      config.modResults.contents = addPrebuildsEnvShim(config.modResults.contents);
-    }
     config.modResults.contents = addNewPodsTarget(
       config.modResults.contents,
       pluginConfig.targetName