+++ packages/expo-brownfield/plugin/src/ios/utils/props.ts
@@ -9,7 +9,6 @@ export const getPluginConfig = (props: PluginProps, config: ExpoConfig): PluginC
     bundleIdentifier: getBundleIdentifier(props, config, targetName),
     targetName,
     buildReactNativeFromSource: props?.buildReactNativeFromSource ?? false,
-    usePrebuilds: props?.usePrebuilds ?? false,
   };
 };