+++ packages/expo-brownfield/CHANGELOG.md
@@ -21,7 +21,7 @@
 - [ios] Fix loading assets in brownfield. ([#44724](https://github.com/expo/expo/pull/44724) by [@gabrieldonadel](https://github.com/gabrieldonadel))
 - [iOS] Support rendering multiple ReactNativeView simultaneously ([#44891](https://github.com/expo/expo/pull/44891) by [@gabrieldonadel](https://github.com/gabrieldonadel))
 - Support registering custom turbo modules from the hosting app ([#44929](https://github.com/expo/expo/pull/44929) by [@gabrieldonadel](https://github.com/gabrieldonadel))
-- [iOS] Support bundling prebuilt Expo module xcframeworks into the Swift Package output via the new `usePrebuilds` plugin prop and `--use-prebuilds` CLI flag. ([@gabrieldonadel](https://github.com/gabrieldonadel))
+- [iOS] `build:ios` now detects prebuilt Expo module xcframeworks in `ios/Pods/` (when `expo-build-properties.ios.usePrecompiledModules` is enabled) and bundles them into the Swift Package output as binary targets behind a single aggregate `.library` product. ([@gabrieldonadel](https://github.com/gabrieldonadel))
 
 ### 🐛 Bug fixes